refactor: move navbar component to layouts folder
