+++ apps/web/src/layouts/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router";
-import NavBar from "../components/NavBar";
+import NavBar from "./components/NavBar";
 import styles from "./Layout.module.css";
 
 export default function Layout() {
