+++ apps/web/src/layouts/components/NavBar.module.css
@@ -15,3 +15,14 @@
   display: flex;
   gap: 10px;
 }
+
+.menuListItem {
+}
+
+.menuListItemLink {
+  padding: 10px;
+  border-radius: 0.25rem;
+  &:hover {
+    background-color: var(--muted);
+  }
+}
